moved Line and User classes to bot/line.py, continuing work on those
